Timezones were not being treated correctly
Fix this by defining an offset from local time
The QDateTimeAxis appears to expect local time,
however, the user_timeseries class uses UTC to
avoid crashes. The offset is subtracted to generate
the correct time in the plot.

Also, fixing a bug that had been in the code for a
while where the timezone was added to the USGS
data instead of subtracted.